sdm660-common: libshims: Respect non-static member function calls.
As stated in
https://www.learncpp.com/cpp-tutorial/8-8-the-hidden-this-pointer/
member functions contains a hidden parameter (a pointer of the current
reference of teh object, aka "this"). This way the code knows in which
context the member function is called. We have to resprect that behavior
into our shims which otherwise crashes or lead to memory corruption.

Change-Id: I68197ebafde8773fd73d7dafb9b2f1bae92ee170
Signed-off-by: pix106 <sbordenave@gmail.com>