sdm660-common: Add (unused) qti mapper@2.0 interface libs for blobs
The blobs link against this interface library because of a fallback
mechanism when mapper@3.0 isn't made available (all our devices should
always provide that, never 2.0). Consequently this results in a silent
failure (not even a linker error appears) when trying to open the GPU
through EGL:

    I AdrenoGLES-0: Driver Path                      : /odm/lib64/egl/libGLESv2_adreno.so
    W libEGL  : eglInitialize(0xb40000743497a2d0) failed (EGL_BAD_ALLOC)
    E libEGL  : call to OpenGL ES API with no current context (logged once per thread)
    F RenderEngine: eglQueryStringImplementationANDROID(EGL_VERSION) failed

Adding this library - which is not explicitly linked/used by any other
bits in the OSS part of the stack - resolves that error.

Note that while the composer has a similar fallback mechanism to 2.0 it
only uses (and links against) the AOSP interface
(`android.hardware.graphics.mapper@2.0`), not the QTI one at
`vendor.qti.hardware.display.mapper@2.0`.

Note^2: All these hidl interfaces are defined as `system_ext`, which is
the desired place for packages (in this case interface
updates/extensions to the AOSP variant) may they ever land upstream.
We have not configured such a partition currently (on super) but instead
rely on it being in `/system/system_ext`. This is not compatible with
GSIs; we just request the vendor variant to be put on `/vendor`
(`.vendor` postfix on the package name).

Signed-off-by: Marijn Suijten <marijns95@gmail.com>
Signed-off-by: Jabiyeff <cebiyevanar@gmail.com>
Signed-off-by: clarencelol <clarencekuiek@icloud.com>
Signed-off-by: pix106 <sbordenave@gmail.com>